Refactoring radio player service
