+++ lib/screens/stations_screen.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:etherly/models/device.dart';
 import 'package:etherly/localization/app_localizations.dart';
 import 'package:etherly/models/station.dart';
-import 'package:etherly/services/radio_player_service.dart';
+import 'package:etherly/services/audio_player_service.dart';
 import 'package:etherly/widgets/screen_header.dart';
 import 'package:etherly/widgets/station_card_item.dart';
 import 'package:etherly/widgets/station_grid_item.dart';
